+++ packages/swapper/src/manager/SwapperManager.ts
@@ -21,18 +21,18 @@ export class SwapperManager {
 
   /**
    *
-   * @param swapperType swapper type {SwapperType|string}
    * @param swapperInstance swapper instance {Swapper}
    * @returns {SwapperManager}
    */
-  addSwapper(swapperType: SwapperType, swapperInstance: Swapper<ChainId>): this {
+  addSwapper(swapperInstance: Swapper<ChainId>): this {
+    validateSwapper(swapperInstance)
+    const swapperType = swapperInstance.getType()
     const swapper = this.swappers.get(swapperType)
     if (swapper)
       throw new SwapError('[addSwapper] - swapper already exists', {
         code: SwapErrorTypes.MANAGER_ERROR,
         details: { swapperType }
       })
-    validateSwapper(swapperInstance)
     this.swappers.set(swapperType, swapperInstance)
     return this
   }
@@ -41,7 +41,6 @@ export class SwapperManager {
    *
    * @param swapperType swapper type {SwapperType|string}
    * @returns {Swapper}
-   * @deprecated this will be removed, currently used in swapper tests
    */
   getSwapper(swapperType: SwapperType): Swapper<ChainId> {
     const swapper = this.swappers.get(swapperType)